windows installer: Do not attempt to pin launcher to taskbar and start menu
The installer was conditionally pinning Zotero's launcher to the taskbar and
start menu. Some of the conditions involved checks for the default browser,
whether the application was already pinned, what version of Windows was
installed, etc.

With this update a desktop and start menu icon get created, but no
attempts are made to pin the launcher to the start menu or taskbar.