Windows uninstaller: Remove all references to HideShortcuts
This patch removes support for the /HideShortcuts command-line
argument in the stub uninstaller. This argument would trigger
the 'HideShortcuts' macro to be called, which would delete any
shortcuts on the desktop.

This behavior was redundant. The 'DeleteShortcuts' macro is always
called, which does the same thing. The 'HideShortcuts' macro requires
elevation to run, and so it can not be used in a per-user uninstall.