Windows uninstaller: Remove all references to ShowShortcuts
This patch removes support for the /ShowShortcuts command-line
argument in the stub installer. This was probably meant to provide
a way to put shortcuts on the desktop and start menu after Firefox
was already installed.

The 'ShowShortcuts' macro requires elevation to run, and so it can
not be used in a per-user uninstall.