Windows installer: do not write to Software\Microsoft\MediaPlayer\ShimInclusionList registry key.
The Windows Media Player application has a list of processes executable files
which are allowed to launch it. Zotero's installer was attempting to add
itself to this list, but was not doing so correctly. The installer was not
initializing a variable to the correct value before writing to the
registry. Instead of Zotero's process name, the strings "true" and "false"
were being added to ShimInclusionList. The uninstaller was then failing to
clean up the registry.