Do not register Zotero as the Start Menu Internet Link.
Windows XP and Windows Vista have a special Start Menu Internet Link in the
start menu. This menu item can be customized by registering an application in
the registry under Software\Clients\StartMenuInternet. This registration is
ignored in Windows 7 and Windows 8.

Zotero's installer was partially filling in this information, and the
uninstaller was failing to remove it.

The uninstaller was also deleting a key under
Software\RegisteredApplications. This key is used for the Windows "Default
Programs" feature, which Zotero is not using. The installer was not writing
values to this key.