Build from XPI instead of git repo
Instead of duplicating the build logic from the XPI build scripts (which
differs between 4.0 and 5.0), and having different logic for building from git
vs. using an existing directory, just use a prebuilt XPI as the basis for all
builds. The version number is pulled from the XPI's install.rdf, with .SA
appended.

This removes the option of symlinking an existing git directory, but we'll soon
be doing code preprocessing (for JSX, etc.) anyway, so it will be easy enough
to build an XPI and Standalone at the same time.

Word processor integration plugins are still built from submodules.