Add Unity launcher file, and script to set its icon
Since 'zotero' is now a shell script that might open in a text editor by
default (as it does on Ubuntu), add a launcher (.desktop) file that can
be clicked on by tarball users. .desktop files stupidly can't use
relative paths, so use a hack to run the executable and add a script
that will update the .desktop file if it's writable with the current
path to the icon. Technically the .desktop file could call the script to
update itself, but it wouldn't take effect until the second launch
anyway and can possibly cause the launcher icon to behave strangely, so
let people run it manually if they care. They should really be using a
package anyway.