Use Firefox 45.0.2esr on Windows and improve .exe metadata
The prebuilt xulrunner-stub.exe from
https://github.com/duanyao/xulrunner-stub only works with Firefox 47 and
lower, so use Firefox 45 ESR until we rebuild it for Firefox 49 (which
we should do to avoid unexpected bugs).

This also replaces xulrunner-stub.exe with a premodified version that
has the icon in place (using ReplaceVistaIcon.exe) and has appropriate
metadata set (using Resource Hacker).