Add support for remote debugging via devtools server
devtools files are packaged in the browser/omni.ja, which we don't
include. This changes fetch_xulrunner.sh to extract them to a directory
in the Firefox package and adds a -t flag to build.sh that copies the
files into the build and adds prefs to allow the debugging server on
port 6100 and disable the prompt. (scripts/dir_build also now has a -t
flag.)

A flag to start the debugging server at runtime is coming in a separate
commit.