Support building for Windows on Linux
Building on a Windows VM is stupidly slow, but we use no
platform-specific tools to not support building non-native builds.
Signing, naturally, has not been tested, but a Windows build, built on
Linux, works on Windows as expected.

This naturally extends to support other cross-platform builds, with
MacOS being the prime contender for building for all platforms.