Switch to compiled Mac stub to fix launch issues under High Sierra
Instead of making the main executable target a shell script that calls
the official Firefox binary with -app, use a modified binary with a
hard-coded relative path to the application.ini file as the main target.

This is based on the FIREFOX_54_0_1_RELEASE tag on mozilla-unified, with
the included patch applied and a custom mozconfig to link against the
10.7 SDK (available from "Xcode 4.3.3 for Lion" in the Apple Developer
Center), which seems to be what the official version does. (Firefox 57
switches to 10.12.) When I built this against the default 10.13 SDK, I
didn't get a unified toolbar and the font was different.

Fixes https://github.com/zotero/zotero/issues/1330