register MIME file types in registry
From my experience with RIS files without extension downloaded by FireFox, they are only loaded in Zotero when using the "Open with... Zotero" option in Firefox, when Zotero is associated in Windows with the .ris files AND there is a registered MIME file type in windows registry, associating application/x-research-info-systems with .ris.  (In my case these RIS files are generated by a bookmarklet and not "caught" by the Zotero browser extension. But loading downloaded ris files should also be possible when no browser extension is installed. Thus I propose this patch to improve the handling of downloaded bibliographic files)

See discussion here: 
https://forums.zotero.org/discussion/62545/use-zotero-for-downloaded-ris