Allow proxy password saving
This causes the "Use Password Manager to remember this password."
checkbox to appear in the proxy authentication dialog. This wasn't
appearing previously (usually, at least [1]) because the proxy XHR
doesn't have an associated window, causing the request to be treated
like it was part of a private browsing session.

Clearing of stored passwords will be handled separately in the client.

[1] https://forums.zotero.org/discussion/comment/353949/#Comment_353949