fx-compat: Initial Firefox 101 compatibility (macOS/Linux only)
In preparation for Firefox 102 ESR

This bundles much more of Firefox than in previous builds -- the entire
browser omni.ja, with Zotero files added, is now included in an 'app'
subdirectory, where the macOS stub and the Linux start script now look
for application.ini. We can likely exclude much more of the browser
omni.ja. This does get the modern browser console in place of the old
error console, though.

TODO:

- Rebuild Windows stub
- Restore or delete various previous customizations that are currently
  disabled in fetch_xulrunner.sh
- General cleanup
- Speed up building by caching Firefox browser omni.ja extraction