Initial Big Sur compatibility
Add modified versions of the XUL bundle and libnss3.dylib from Firefox
60 to fix dlopen() (required for both js-ctypes and OpenGL support) and
an issue where menus and popups appear blank. Hopefully these changes
won't cause problems in older versions of macOS. There are some other
minor graphical issues in Big Sur, some of which are fixable via CSS and
some that we may be able to backport fixes for from current Firefox once
Mozilla addresses them.

For some reason, using a freshly built version of the stub causes the
pane-switching animation in the preferences to be really glitchy on
Catalina, whereas that doesn't happen with the previous version of the
stub. Unfortunately and strangely, it seems to happen on Big Sur even
with the old stub.

(The modified files are 100 MB, so they're not included here, but
they'll be available in the distributed Zotero.app. They can also be
built from the mozilla-unified FIREFOX_60_9_0esr_RELEASE tag with the
included patch and mozconfig file.)