Fix scaling issue when switching between monitors on Big Sur
https://forums.zotero.org/discussion/comment/363312/#Comment_363312

On Big Sur, the backing scale change doesn't seem to be properly
reported, causing the window contents to appear in 1/4 of the window
when switching from Retina to non-Retina (and get really big going the
other direction). Resizing the window fixes the problem, though, so as a
hack when we detect a screen change we just shrink the window height by
a pixel and increase it again. windowDidChangeScreen is called at odd
times, including startup, and can be called for zero-size windows, but
if we limit our fix to non-zero-size windows it seems to work without
causing problems.