Use `ZOTERO_PROFILE` environment variable for profile in `build_and_run`
The previous advice [1] was to copy this script somewhere in your path,
but since that would miss further changes to the file (such as the
addition of the `-q` flag), we should recommend simply calling it from
another helper script in the path or a shell alias.

To make things more flexible, this removes the hard-coded `PROFILE`
variable and expects that to be set in a `ZOTERO_PROFILE` environment
variable if desired. E.g., I now use this helper script in my path:

```
 #!/bin/bash -e
 export ZOTERO_PROFILE="Dev"
 ~/zotero-standalone-build/scripts/build_and_run "$@"
```

[1] https://www.zotero.org/support/dev/client_coding/building_the_standalone_client#helper_script