Fix Word integration from source builds on Apple Silicon
From what I can gather [1], all apps on Apple Silicon need to be signed,
but macOS will automatically ad-hoc sign...something. (The linked page
says if you've "compiled the executable yourself", but it doesn't seem
to be limited to that.) In any case, the automatic signing doesn't
happen for the Word dylib, so we need to add a `codesign` call to sign
that file, and then it works.

I put the signing code in a separate file, in case we discover other
things that need to be explicitly signed, and because someone making a
custom build for their own usage might want to sign the full app bundle.

[1] https://wiki.lazarus.freepascal.org/Code_Signing_for_macOS#Ad_hoc_signing