Revert "Remove textActions.ftl from Firefox build"
This reverts commit 4b9729b6b638ffadc0b2df358c85e9e9adb90c60.

Without this file the Edit menu and most of its menu options are missing
labels, at least on macOS. Menu option labels are missing even if
`MozXULElement.insertFTLIfNeeded("toolkit/global/textActions.ftl")` is
changed to point to `mozilla/textActions.ftl` in our editMenuOverlay.js.
We'll need to figure out how to use our versions, but for now we'll have
to live with the extra console output.